Refactoring of type_checking mod.
